add treeBox deploy script and rename treeBoxV2 to treeBox
